misc/HtmlTestSuite/results/test-0162.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS Intrinsic &amp; Extrinsic Sizing Test: parsing height with valid values</title>
<link href="https://drafts.csswg.org/css-sizing-3/#propdef-height" rel="help"/>
<meta content="height supports the full grammar 'auto | &lt;length-percentage&gt; | min-content | max-content | fit-content(&lt;length-percentage&gt;)'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
</head>
<body>
<script>
test_valid_value("height", "auto");
test_valid_value("height", "min-content");
test_valid_value("height", "max-content");

test_valid_value("height", "0", "0px");
test_valid_value("height", "10%");
test_valid_value("height", "0.5em");
test_valid_value("height", "calc(10% - 0.5em)");

test_valid_value("height", "fit-content(10%)");
test_valid_value("height", "fit-content(0.5em)");
test_valid_value("height", "fit-content(calc(10% - 0.5em))");
</script>
</body>
</html>
